% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorplaner.R
\docType{package}
\name{colorplaner}
\alias{colorplaner}
\alias{colorplaner-package}
\title{colorplaner: ggplot2 Extenstion to Visualize Two Variables Per Color
Aesthetic through Color Space Projection}
\description{
Add additional dimensionality to
visualizations by using the color and/or fill aesthetics to convey the values
of two continous variables each. By projecting variable values onto YUV
colorspace, a scale is created that allows viewers to intuitively determine
the values of both variables from the single displayed color. Includes two
new scales and a new guide for ggplot2.
}

