% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_colorplane.R
\name{guide_gengrob.colorplane}
\alias{guide_gengrob.colorplane}
\title{Colorplane method for guide_gengrob}
\usage{
guide_gengrob.colorplane(guide, theme)
}
\arguments{
\item{guide}{Object of class "colorplane" generate by
\code{\link{guide_colorplane}}}

\item{theme}{Plot theme object}
}
\description{
Called by ggplot2 engine to create graphical objects for the guide.
}
\seealso{
\code{\link{scale_color_colorplane}}
}
\keyword{internal}

