% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_colorplane.R
\name{guide_geom.colorplane}
\alias{guide_geom.colorplane}
\title{Colorplane method for guide_geom}
\usage{
\method{guide_geom}{colorplane}(guide, ...)
}
\arguments{
\item{guide}{Object of class "colorplane" generate by
\code{\link{guide_colorplane}}}

\item{...}{Not used}
}
\description{
Called by ggplot2 engine. Takes no action as this guide is not geom-based.
}
\keyword{internal}
