% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_colorplane.R
\name{guide_merge.colorplane}
\alias{guide_merge.colorplane}
\title{Colorplane method for guide_merge}
\usage{
\method{guide_merge}{colorplane}(guide, new_guide)
}
\arguments{
\item{guide}{Object of class "colorplane" generate by
\code{\link{guide_colorplane}}}

\item{new_guide}{New guide object}
}
\description{
Called by ggplot2 engine. Ignores attempts to merge guides.
}
\keyword{internal}
