% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_colorplane.R
\name{guide_train.colorplane}
\alias{guide_train.colorplane}
\title{Colorplane method for guide_train}
\usage{
\method{guide_train}{colorplane}(guide, scale, aesthetic = NULL)
}
\arguments{
\item{guide}{Object of class "colorplane" generate by
\code{\link{guide_colorplane}}}

\item{scale}{ggproto object instance of \code{\link{ScaleColorPlane}}}
}
\description{
Called by ggplot2 engine when building the plot.
}
\keyword{internal}
