% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleColorPlane.R
\docType{data}
\name{ScaleColorPlane}
\alias{ScaleColorPlane}
\title{Color Plane Scale ggproto Object}
\format{An object of class \code{ScaleColorPlane} (inherits from \code{ScaleContinuous}, \code{Scale}, \code{ggproto}) of length 16.}
\usage{
ScaleColorPlane
}
\description{
This ggproto object inherits from \code{\link[ggplot2]{ScaleContinuous}} and
implements methods and default values needed for color plane scale instances.
See \code{\link{scale_color_colorplane}} for usage.
}
\keyword{datasets}

