% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorplaner.R
\name{colorplaner}
\alias{colorplaner}
\title{colorplaner: ggplot2 Extension to Visualize Two Variables Per Color
Aesthetic through Color Space Projection}
\description{
Add additional dimensionality to visualizations by using the color and/or
fill aesthetics to convey the values of two continuous variables each. By
projecting variable values onto YUV color space, a scale is created that
allows viewers to intuitively determine the values of both variables from the
single displayed color. Includes two new scales and a new guide for ggplot2.
See \code{\link{scale_color_colorplane}} for usage.
}
\section{Warning Message About Ignoring Unknown Aesthetics}{

Layers now produce a warning message when unrecognized aesthetics are found
but have no mechanism for notifying them of aesthetics handled by scales.
The warning can be avoided by mapping \code{color2}/\code{fill2} at the plot
level (i.e. in the initial \code{ggplot()} statement). If you want to avoid
colorplane mapping on all layers, map \code{color}/\code{fill} only on the
layers you want, as in the example below.
}
\examples{
library(ggplot2)
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width,
  colour2 = Petal.Width)) +
 geom_point(aes(colour = Petal.Length)) +
 geom_line(aes(linetype = Species)) +
 scale_color_colorplane()


}

