% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_colorplane.R
\name{guide_colorplane}
\alias{guide_colorplane}
\alias{guide_colourplane}
\title{Add Guide for Colorplane}
\usage{
guide_colorplane(title = waiver(), title.position = NULL,
  title.theme = NULL, title.hjust = 0.5, title.vjust = NULL,
  axis_title = waiver(), axis_title.position = NULL,
  axis_title.theme = NULL, axis_title.hjust = NULL,
  axis_title.vjust = NULL, axis_title_y = waiver(),
  axis_title_y.position = NULL, axis_title_y.theme = NULL,
  axis_title_y.hjust = NULL, axis_title_y.vjust = NULL, label = TRUE,
  label.position = NULL, label.theme = NULL, label.hjust = NULL,
  label.vjust = NULL, label_y.position = NULL, label_y.theme = NULL,
  label_y.hjust = NULL, label_y.vjust = NULL, planewidth = NULL,
  planeheight = NULL, nbin = 20, ticks = TRUE, default.unit = "line",
  order = 0, ...)

guide_colourplane(title = waiver(), title.position = NULL,
  title.theme = NULL, title.hjust = 0.5, title.vjust = NULL,
  axis_title = waiver(), axis_title.position = NULL,
  axis_title.theme = NULL, axis_title.hjust = NULL,
  axis_title.vjust = NULL, axis_title_y = waiver(),
  axis_title_y.position = NULL, axis_title_y.theme = NULL,
  axis_title_y.hjust = NULL, axis_title_y.vjust = NULL, label = TRUE,
  label.position = NULL, label.theme = NULL, label.hjust = NULL,
  label.vjust = NULL, label_y.position = NULL, label_y.theme = NULL,
  label_y.hjust = NULL, label_y.vjust = NULL, planewidth = NULL,
  planeheight = NULL, nbin = 20, ticks = TRUE, default.unit = "line",
  order = 0, ...)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link{waiver}}), the name of the scale object or the name
specified in \code{\link{labs}} is used for the title.}

\item{title.position}{A character string indicating the position of a
 title. One of "top" (default for a vertical guide), "bottom", "left"
(default for a horizontal guide), or "right."}

\item{title.theme}{A theme object for rendering the title text. Usually the
object of \code{\link{element_text}} is expected. By default, the theme is
specified by \code{legend.title} in \code{\link{theme}} or theme.}

\item{title.hjust}{A number specifying horizontal justification of the
title text.}

\item{title.vjust}{A number specifying vertical justification of the title
text.}

\item{axis_title, axis_title_y}{Character strings or expressions inidicating
the horizontal and vertical axis titles in the guide, respectively. If
\code{NULL}, the title is not shown. By default (\link[ggplot2]{waiver}),
the name of the scale or the name of the variable mapped to the aesthetic.}

\item{axis_title.position, axis_title_y.position}{Character strings indicating
the positions of the axis titles. \code{axis_title.position}: one of "top"
or "bottom" (default). \code{axis_title_y.position}: one of "left"
(default) or "right". Not yet implemented.}

\item{axis_title.theme, axis_title_y.theme}{Theme objects for rendering the
axis title text. Typically an \code{\link[ggplot2]{element_text}} object.
When \code{NULL}, defaults to settings for \code{axis.title.x} and
\code{axis.title.y} in the plot theme.}

\item{axis_title.hjust, axis_title_y.hjust}{Numbers specifying horizontal
justification of the axis title text. When \code{NULL}, defaults to
\code{axis_title.theme} or \code{axis.title.x} in plot theme.}

\item{axis_title.vjust, axis_title_y.vjust}{Numbers specifing veritcal
justification of the axis title text. When \code{NULL}, defaults to
\code{axis_title*.theme} if set or \code{axis.title.*} in plot theme.}

\item{label}{logical. If \code{TRUE} then the labels are drawn. If
\code{FALSE} then the labels are invisible.}

\item{label.position, label_y.position}{Character strings indicating the
positions of axis labels. For \code{label.position}, "top" or "bottom"
(default). For \code{label_y.position}, "left" (default) or "right". Not
yet implemented.}

\item{label.theme, label_y.theme}{Theme objects for rendering axis label text.
Usually the object of \code{\link[ggplot2]{element_text}} is expected. By
default, the theme is specified by \code{axis.text.*} in the plot theme.}

\item{label.hjust, label_y.hjust}{Numerics specifying horizontal justification
of the axis label text. Defauls to value in \code{label.theme} /
\code{label_y.theme} if set or \code{axis.text.*} in the plot theme.}

\item{label.vjust, label_y.vjust}{Numerics specifying vertical justification
of the axis label text. Defauls to value in \code{label.theme} /
\code{label_y.theme} if set or \code{axis.text.*} in the plot theme.}

\item{planewidth, planeheight}{Numeric or \code{\link[grid]{unit}} objects
specifying the width and height of the colorplane. Default values are 5
times the \code{legend.key.width/height} or \code{legend.key.size} in the
plot theme.}

\item{nbin}{Number specifying how many color pixels are generated for each
dimension of the colorplane. Higher numbers increase guide color accuracy
(especially for larger sized guides) at the expense of speed.}

\item{ticks}{A logical specifying if tick marks on colorbar should be
visible.}

\item{default.unit}{A character string indicating \code{\link[grid]{unit}}
for \code{barwidth} and \code{barheight}.}

\item{order}{positive integer less that 99 that specifies the order of
this guide among multiple guides. This controls the order in which
multiple guides are displayed, not the contents of the guide itself.
If 0 (default), the order is determined by a secret algorithm.}

\item{...}{ignored.}
}
\description{
Generates a guide to explain the colors plotted via
\code{\link{scale_color_colorplane}} and \code{\link{scale_fill_colorplane}}.
}
\details{
The guide is similar to \code{\link[ggplot2]{guide_colorbar}}, but as a plane
of colors with ticks and labels for both variables in the scale.
}
\examples{
if(requireNamespace("mapproj")) {
  crimes <- data.frame(state = tolower(rownames(USArrests)), USArrests)
  states_map <- map_data("state")
  ggplot(crimes,
         aes(map_id = state, fill = Murder, fill2 = UrbanPop)) +
    geom_map(map = states_map) +
    scale_fill_colorplane() +
    expand_limits(x = states_map$long, y = states_map$lat) +
    coord_map() +
    guides(fill = guide_colorplane("My Title", axis_title = "Murder Rate",
    axis_title_y = "Urban Population \%"))
 }
}

